backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docx2txt>=0.9",
    "fastapi[all]>=0.116.1",
    "fastembed>=0.7.3",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.7",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.0",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
    "unstructured>=0.18.15",
    "uvicorn>=0.35.0",
    "tomli"
]
